backend/requirements.txt
# Core
Django==5.2.0
gunicorn==23.0.0

# API
djangorestframework==3.16.1
django-filter==25.2
django-cors-headers==4.4.0
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.27.2

# DB (PostgreSQL)
dj-database-url==2.2.0
psycopg2-binary==2.9.9

# Config / env
python-decouple==3.8
python-dotenv==1.0.1

# Media / images
pillow==10.4.0

# Payments
stripe==14.1.0

requests==2.32.3
urllib3==2.2.3
certifi==2026.1.4
charset-normalizer==3.3.2
idna==3.10

# Django dependency
asgiref==3.11.0
sqlparse==0.5.1
typing_extensions==4.15.0

whitenoise==6.11.0

cloudinary==1.41.0
django-cloudinary-storage==0.3.0